refactor: create utility function to convert FileList into FileEntries
